Migrate type enums into viewer.
